Add files via upload
Insti-bot has the capability to answer as many questions as the user asks it. Insti-bot uses Microsoft Azure Services to understand the queries raised by the student or a parent, using it’s QnA Maker services and a Bot services. The QnA Maker allows the institution to create a Knowledge Base of its own. Web App bot enables the institution to bring their bot live and reach as many people as possible to improve their business. Also, using Microsoft’s Azure services will allow the institution, “pay- as- you- go” option that reduces the cost for maintaining the bot.